HistoryToken.setSaveValue abstract
- Extracted into sub-classes setSaveValue
