+++ src/main/java/walkingkooka/spreadsheet/dominokit/history/HistoryToken.java
@@ -3986,344 +3986,9 @@ public abstract class HistoryToken implements HasUrlFragment {
         return this.setSaveStringValue("");
     }
 
-    /**
-     * if possible creates a save history token, otherwise returns this. For {@link HistoryToken} that do not support saving,
-     * the value is ignored and this is returned.
-     */
-    public final HistoryToken setSaveValue(final Optional<?> value) {
-        Objects.requireNonNull(value, "value");
+    abstract public HistoryToken setSaveValue(final Optional<?> value);
 
-        final Object valueOrNull = value.orElse(null);
-
-        HistoryToken historyToken = null;
-
-        if (this instanceof PluginHistoryToken) {
-            historyToken = this.cast(PluginHistoryToken.class)
-                .setSaveValuePlugin(value);
-        }
-
-        if (this instanceof SpreadsheetIdHistoryToken) {
-            final SpreadsheetId id = this.cast(SpreadsheetIdHistoryToken.class)
-                .id();
-
-            if (this instanceof SpreadsheetListRenameHistoryToken) {
-                if (value.isPresent()) {
-                    historyToken = HistoryToken.spreadsheetListRenameSave(
-                        id,
-                        (SpreadsheetName) value.get()
-                    );
-                } else {
-                    historyToken = HistoryToken.spreadsheetListRenameSelect(id);
-                }
-            }
-
-            if (this instanceof SpreadsheetNameHistoryToken) {
-                final SpreadsheetName name = this.cast(SpreadsheetNameHistoryToken.class)
-                    .name();
-
-                if (this instanceof SpreadsheetRenameHistoryToken) {
-                    if (value.isPresent()) {
-                        historyToken = HistoryToken.spreadsheetRenameSave(
-                            id,
-                            name,
-                            (SpreadsheetName) value.get()
-                        );
-                    } else {
-                        historyToken = HistoryToken.spreadsheetRenameSelect(
-                            id,
-                            name
-                        );
-                    }
-                }
-
-                if (this instanceof SpreadsheetAnchoredSelectionHistoryToken) {
-                    final AnchoredSpreadsheetSelection spreadsheetSelection = this.cast(SpreadsheetAnchoredSelectionHistoryToken.class)
-                        .anchoredSelection();
-
-                    if (this instanceof SpreadsheetCellHistoryToken) {
-                        if (this instanceof SpreadsheetCellSelectHistoryToken || this instanceof SpreadsheetCellSaveHistoryToken) {
-                            historyToken = this.cast(SpreadsheetCellHistoryToken.class)
-                                .setSaveValueCell(valueOrNull);
-                        }
-
-                        if (this instanceof SpreadsheetCellDateTimeSymbolsHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof DateTimeSymbols) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    DateTimeSymbols.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellDateTimeSymbolsSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellDecimalNumberSymbolsHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof DecimalNumberSymbols) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    DecimalNumberSymbols.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellDecimalNumberSymbolsSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellFormatterHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof SpreadsheetFormatterSelector) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetFormatterSelector.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellFormatterSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellFormulaHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof String) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    String.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellFormulaSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                CharSequences.nullToEmpty((String) valueOrNull)
-                                    .toString()
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellLabelHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof SpreadsheetLabelName) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetLabelName.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellLabelSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                (SpreadsheetLabelName) valueOrNull
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellLocaleHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof Locale) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    Locale.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellLocaleSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellParserHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof SpreadsheetParserSelector) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetParserSelector.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellParserSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellSortHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof SpreadsheetColumnOrRowSpreadsheetComparatorNamesList) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetColumnOrRowSpreadsheetComparatorNamesList.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellSortSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                (SpreadsheetColumnOrRowSpreadsheetComparatorNamesList) valueOrNull
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellStyleHistoryToken) {
-                            historyToken = HistoryToken.cellStyleSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                this.cast(SpreadsheetCellStyleHistoryToken.class)
-                                    .propertyName(),
-                                value
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellValidatorHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof ValidatorSelector) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    ValidatorSelector.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellValidatorSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellValueHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof String) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    String.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellValueSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                this.cast(SpreadsheetCellValueHistoryToken.class)
-                                    .valueType()
-                                    .get(),
-                                CharSequences.nullToEmpty((String) valueOrNull)
-                                    .toString()
-                            );
-                        }
-
-                        if (this instanceof SpreadsheetCellValueTypeHistoryToken) {
-                            if (null != valueOrNull && false == valueOrNull instanceof ValueTypeName) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    ValueTypeName.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.cellValueTypeSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(value)
-                            );
-                        }
-                    }
-
-                    if (this instanceof SpreadsheetColumnSortHistoryToken) {
-                        if (null != valueOrNull) {
-                            if (false == valueOrNull instanceof SpreadsheetColumnOrRowSpreadsheetComparatorNamesList) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetColumnOrRowSpreadsheetComparatorNamesList.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.columnSortSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(valueOrNull)
-                            );
-                        }
-                    }
-
-                    if (this instanceof SpreadsheetRowSortHistoryToken) {
-                        if (null != valueOrNull) {
-                            if (false == valueOrNull instanceof SpreadsheetColumnOrRowSpreadsheetComparatorNamesList) {
-                                this.reportInvalidSaveValue(
-                                    valueOrNull,
-                                    SpreadsheetColumnOrRowSpreadsheetComparatorNamesList.class
-                                );
-                            }
-
-                            historyToken = HistoryToken.rowSortSave(
-                                id,
-                                name,
-                                spreadsheetSelection,
-                                Cast.to(valueOrNull)
-                            );
-                        }
-                    }
-                }
-
-                if (this instanceof SpreadsheetLabelMappingSelectHistoryToken || this instanceof SpreadsheetLabelMappingSaveHistoryToken) {
-                    if (null != valueOrNull && false == valueOrNull instanceof SpreadsheetExpressionReference) {
-                        this.reportInvalidSaveValue(
-                            valueOrNull,
-                            SpreadsheetExpressionReference.class
-                        );
-                    }
-
-                    if (null != valueOrNull) {
-                        historyToken = HistoryToken.labelMappingSave(
-                            id,
-                            name,
-                            this.cast(SpreadsheetLabelMappingHistoryToken.class)
-                                .labelName()
-                                .get()
-                                .setLabelMappingReference((SpreadsheetExpressionReference) valueOrNull)
-                        );
-                    }
-                }
-
-                if (this instanceof SpreadsheetMetadataHistoryToken) {
-                    if (this instanceof SpreadsheetMetadataPropertySelectHistoryToken || this instanceof SpreadsheetMetadataPropertySaveHistoryToken) {
-                        historyToken = HistoryToken.metadataPropertySave(
-                            id,
-                            name,
-                            this.cast(SpreadsheetMetadataPropertyHistoryToken.class)
-                                .propertyName,
-                            value
-                        );
-                    }
-
-                    if (this instanceof SpreadsheetMetadataPropertyStyleHistoryToken) {
-                        historyToken = HistoryToken.metadataPropertyStyleSave(
-                            id,
-                            name,
-                            this.cast(SpreadsheetMetadataPropertyStyleHistoryToken.class)
-                                .stylePropertyName(),
-                            value
-                        );
-                    }
-                }
-            }
-        }
-
-        return elseIfDifferent(historyToken);
-    }
-
-    private void reportInvalidSaveValue(final Object value,
+    final void reportInvalidSaveValue(final Object value,
                                         final Class<?> expected) {
         // Invalid value: got String expected Integer
 
@@ -5145,7 +4810,7 @@ public abstract class HistoryToken implements HasUrlFragment {
         return historyToken;
     }
 
-    private HistoryToken elseIfDifferent(final HistoryToken historyToken) {
+    final HistoryToken elseIfDifferent(final HistoryToken historyToken) {
         return null == historyToken || this.equals(historyToken) ?
             this :
             historyToken;
