+++ src/main/java/walkingkooka/spreadsheet/dominokit/history/SpreadsheetReloadHistoryToken.java
@@ -23,6 +23,9 @@ import walkingkooka.spreadsheet.SpreadsheetName;
 import walkingkooka.spreadsheet.dominokit.AppContext;
 import walkingkooka.text.cursor.TextCursor;
 
+import java.util.Objects;
+import java.util.Optional;
+
 /**
  * A token that represents a reload spreadsheet action.
  * <pre>
@@ -70,6 +73,13 @@ public final class SpreadsheetReloadHistoryToken extends SpreadsheetNameHistoryT
             id,
             name
         );
+    }
+
+    @Override
+    public HistoryToken setSaveValue(final Optional<?> value) {
+        Objects.requireNonNull(value, "value");
+
+        return this;
     }
 
     @Override
